pyproject.toml
[project]
name = "healthpilot"
version = "0.1.0"
description = "Multi-agent AI health assistant powered by LangGraph and Claude"
readme = "README.md"
requires-python = ">=3.11,<3.14"
license = {text = "MIT"}

dependencies = [
    # Agent framework
    "langgraph>=0.2",
    "langchain>=0.3",
    "langchain-anthropic>=0.3",
    "langchain-openai>=0.2",
    "langchain-chroma>=0.2",
    "langchain-text-splitters>=0.3",
    "langsmith>=0.2",

    # Vector store
    "chromadb>=0.5",

    # Data & analytics
    "pandas>=2.2",
    "numpy>=1.26",
    "scipy>=1.14",
    "plotly>=5.22",
    "statsmodels>=0.14",

    # Frontend
    "streamlit>=1.38",

    # Google Calendar
    "google-api-python-client>=2.150",
    "google-auth-oauthlib>=1.2",
    "google-auth-httplib2>=0.2",
    "icalendar>=5.0",

    # Configuration
    "pydantic>=2.9",
    "pydantic-settings>=2.5",
    "python-dotenv>=1.0",

    # HTTP & utilities
    "httpx>=0.27",
    "tenacity>=9.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "mypy>=1.13",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.ruff.lint.isort]
known-first-party = ["agents", "analytics", "rag", "config"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.hatch.build.targets.wheel]
packages = ["agents", "analytics", "rag"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
